refactor: moving towards one class service for doing zk-proofs
