+++ src/background/services/zkProof/protocols/__tests__/zkProofSemaphore.test.ts
